requirements/local.txt
-r base.txt # includes the base.txt requirements file

django-debug-toolbar==3.4.0
django-extensions==3.2.3
Werkzeug==3.0.3
factory-boy==3.3.0

# linters and formatters
black==24.4.2
